backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
# Optional (only during development)
httpx            # for testing async routes
